add test

+++ cmd/pd-server/main.go
@@ -78,7 +78,7 @@ func main() {
 	if err != nil {
 		log.Fatal("initialize logger error", errs.ZapError(err))
 	}
-	logutil.SetRedactLog(cfg.EnableRedactLog == 1)
+	logutil.SetRedactLog(cfg.EnableRedactLog == int32(1))
 
 	server.LogPDInfo()
